+++ lib/src/bond.dart
@@ -8,11 +8,11 @@ import 'package:scoped/src/scope.dart';
 class Bond<T extends Fluid> extends StatelessWidget {
   const Bond({this.builder, this.fluid});
 
-  final FluidBuilderFn<T> builder;
+  final FluidBuilderDelegate<T> builder;
   final T fluid;
 
   Widget build(BuildContext context) {
     return FluidBuilder(
-        fluid: fluid ?? Scope.get<T>(context), builder: builder);
+        fluid: fluid ?? Scope.of(context).store.get<T>(), builder: builder);
   }
 }