+++ lib/src/fluid.dart
@@ -1,16 +1,17 @@
-import 'package:flutter/widgets.dart';
-import 'package:scoped/src/binder.dart';
+import 'dart:async';
 
-typedef Widget FluidBuilderFn<T extends Fluid>(
+import 'package:flutter/widgets.dart';
+
+typedef Widget FluidBuilderDelegate<T extends Fluid>(
     BuildContext context, T fluid);
 
-typedef Widget FluidBuilderWrapFn<T extends Fluid>(
+typedef Widget NotifyBuilderChildDelegate<T extends Fluid>(
     BuildContext context, T fluid, Widget child);
 
 ///Subscribes to a Fluid model to build for each notification
 class FluidBuilder<T extends Fluid> extends StatefulWidget {
   final T fluid;
-  final FluidBuilderFn<T> builder;
+  final FluidBuilderDelegate<T> builder;
 
   FluidBuilder({this.fluid, this.builder});
 
@@ -18,23 +19,23 @@ class FluidBuilder<T extends Fluid> extends StatefulWidget {
 }
 
 class _FluidBuilderState<T extends Fluid> extends State<FluidBuilder<T>> {
-  void _onChange() =>  setState(() {});
+  void _onChange() => setState(() {});
 
   void initState() {
     super.initState();
-    widget.fluid.bind(_onChange);
+    widget.fluid.listen(_onChange);
   }
 
   void dispose() {
-    widget.fluid.free(_onChange);
+    widget.fluid.forget(_onChange);
     super.dispose();
   }
 
   void didUpdateWidget(FluidBuilder oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.fluid != oldWidget.fluid) {
-      oldWidget.fluid.free(_onChange);
-      widget.fluid.bind(_onChange);
+      oldWidget.fluid.forget(_onChange);
+      widget.fluid.listen(_onChange);
     }
   }
 
@@ -43,12 +44,23 @@ class _FluidBuilderState<T extends Fluid> extends State<FluidBuilder<T>> {
 
 ///Listenable implementation for models
 abstract class Fluid {
-  Binder _binder = Binder();
+  final Set<VoidCallback> _bonds = Set<VoidCallback>();
 
-  void bind(VoidCallback bond) => _binder.bind(bond);
+  void listen(VoidCallback bond) => _bonds.add(bond);
 
-  void free(VoidCallback bond) => _binder.free(bond);
+  void forget(VoidCallback bond) => _bonds.remove(bond);
+
+  int _currentVersion = 0;
+  int _targetVersion = 0;
 
   @protected
-  notify() => _binder.notify();
+  void notify() {
+    if (_targetVersion == _currentVersion) {
+      _targetVersion++;
+      scheduleMicrotask(() {
+        _targetVersion = ++_currentVersion;
+        _bonds.toList().forEach((VoidCallback bond) => bond());
+      });
+    }
+  }
 }