+++ lib/src/scope.dart
@@ -4,14 +4,13 @@ import 'package:scoped/src/store.dart';
 ///Store provider widget
 class Scope extends InheritedWidget {
   final Store store;
-  Scope({Store store, Widget child, Key key})
-      : store = store ?? Store(),
-        super(child: child, key: key);
+  Scope({@required this.store, Widget child, Key key})
+      : super(child: child, key: key);
 
   bool updateShouldNotify(InheritedWidget oldWidget) => true;
 
-  static Scope of<T>(BuildContext context) =>
-      context.inheritFromWidgetOfExactType(Scope) as Scope;
+  static Scope of(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<Scope>();
 
   static T get<T>(BuildContext context) => of(context).store.get<T>();
   static call<T>(BuildContext context) => of(context).store.call<T>();